Fix error message when linking through a non-relation field.
When a ModelResource's `fields` include something that looks like a model
relation (such as `model__field`), but `model` is a plain field instead of
a relation, we used to crash with an error message that did nothing to
identify the bad field specification. This patch improves the error message
on failure.